backend/requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
httpx>=0.20.0
Jinja2>=3.0.0
